Split AxSpan to VSpan and HSpan classes.
